% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_performance.R
\name{evaluate_performance}
\alias{evaluate_performance}
\title{Evaluate Model Performance (ROC and Precision-Recall)}
\usage{
evaluate_performance(predicted_probs, actual_classes, positive_label)
}
\arguments{
\item{predicted_probs}{Numeric vector of predicted probabilities for the positive class.}

\item{actual_classes}{Factor or character vector of the actual true labels.}

\item{positive_label}{Character string. The target class you want to evaluate.}
}
\value{
An object of class `quickSentiment_eval`, which is a list containing the following metrics:
  \item{target_class}{Character. The specific positive label used for the evaluation.}
  \item{auc_roc}{Numeric. The Area Under the Receiver Operating Characteristic curve.}
  \item{best_threshold_roc}{Numeric. The optimal probability threshold that maximizes Youden's J statistic.}
  \item{auc_pr}{Numeric. The Area Under the Precision-Recall curve.}
  \item{best_threshold_pr}{Numeric. The probability threshold that maximizes the F1-Score.}
  \item{accuracy_at_best}{Numeric. The overall accuracy of the model if `best_threshold_pr` is applied.}
  \item{roc}{S3 object containing the ROC curve data and metrics.}
  \item{prc}{S3 object containing the Precision-Recall curve data and metrics.}
  \item{threshold_summary}{A data frame summarizing Accuracy, Precision, Recall, and F1 at 0.1 threshold increments.}
}
\description{
Evaluate Model Performance (ROC and Precision-Recall)
}
