% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.quickSentiment_eval.R
\name{plot.quickSentiment_prc}
\alias{plot.quickSentiment_prc}
\title{Plot Precision-Recall Curve}
\usage{
\method{plot}{quickSentiment_prc}(x, ...)
}
\arguments{
\item{x}{An object of class `quickSentiment_prc`.}

\item{...}{Additional graphical parameters.}
}
\description{
Plot Precision-Recall Curve
}
