% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.quickSentiment_eval.R
\name{plot.quickSentiment_roc}
\alias{plot.quickSentiment_roc}
\title{Plot ROC Curve}
\usage{
\method{plot}{quickSentiment_roc}(x, ...)
}
\arguments{
\item{x}{An object of class `quickSentiment_roc`.}

\item{...}{Additional graphical parameters.}
}
\description{
Plot ROC Curve
}
