% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.quickSentiment_eval.R
\name{print.quickSentiment_eval}
\alias{print.quickSentiment_eval}
\title{Print quickSentiment Evaluation Results}
\usage{
\method{print}{quickSentiment_eval}(x, ...)
}
\arguments{
\item{x}{An object of class `quickSentiment_eval`.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print quickSentiment Evaluation Results
}
