% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\docType{data}
\name{qs_negations}
\alias{qs_negations}
\title{Standard Negation Words for Sentiment Analysis}
\format{
An object of class \code{character} of length 25.
}
\usage{
qs_negations
}
\description{
A character vector of 25 common negation words. These words are automatically
protected by the \code{\link{pre_process}} function when \code{retain_negations = TRUE}
to prevent standard stopword lists from destroying sentiment polarity.
}
\keyword{datasets}
