% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOW.R
\name{BOW_train}
\alias{BOW_train}
\title{Train a Bag-of-Words Model}
\usage{
BOW_train(doc, weighting_scheme = "count", ngram_size = 1)
}
\arguments{
\item{doc}{A character vector of documents to be processed.}

\item{weighting_scheme}{A string specifying the weighting to apply. One of
"count" (default), "tf", or "tfidf".}

\item{ngram_size}{An integer specifying the maximum n-gram size. For example,
`ngram_size = 2` will create unigrams and bigrams. Defaults to 1.}
}
\value{
A quanteda `dfm` object to be used as a template for new data.
}
\description{
Train a Bag-of-Words Model
}
