% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit_model}
\alias{logit_model}
\title{Train a Regularized Logistic Regression Model using glmnet}
\usage{
logit_model(train_vectorized, Y, test_vectorized)
}
\arguments{
\item{train_vectorized}{The training feature matrix (e.g., a `dfm` from quanteda).
This should be a sparse matrix.}

\item{Y}{The response variable for the training set. Should be a factor for
classification.}

\item{test_vectorized}{The test feature matrix, which must have the same
features as `train_vectorized`.}
}
\value{
A list containing two elements:
  \item{pred}{A vector of class predictions for the test set.}
  \item{model}{The final, trained `cv.glmnet` model object.}
}
\description{
This function trains a logistic regression model using Lasso regularization
via the glmnet package. It uses cross-validation to automatically find the
optimal regularization strength (lambda).
}
