% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prediction}
\alias{prediction}
\title{Predict Sentiment on New Data Using a Saved Pipeline Artifact}
\usage{
prediction(pipeline_object, df, text_column)
}
\arguments{
\item{pipeline_object}{A list object returned by the main `pipeline()` function.
It must contain the trained model, DFM template, preprocessing function,
and n-gram settings.}

\item{new_raw_text_vector}{A character vector of new, raw documents to be scored.}
}
\value{
A vector of class predictions for the new data.
}
\description{
This is a generic prediction function that handles different model types
and ensures consistent preprocessing and vectorization for new, unseen text.
}
